Update issue 794
